1. Change balance field's data type from int32 to decimal 2. Small modification to reduce LoC

+++ hMoney/Account.cs
@@ -15,7 +15,7 @@ namespace hMoney
         public String Status;
         public String Notes;
         public String HeldAt;
-        public int Balance;     // This field is calculated dynamically, not existed in database
-        public int InitialBal;
+        public decimal Balance;     // This field is calculated dynamically, not existed in database
+        public decimal InitialBal;
     }
 }
